requirements.txt
flask
flask-cors
python-escpos